requirements.txt
fastapi
uvicorn
pytest
requests
python-multipart
PyJWT
passlib[bcrypt]
